requirements.txt
tensorflow-probability==0.24.0
tensorflow==2.10.1
# cuda==12.3
# pip3 install --user "tensorflow[and-cuda]"
pillow==10.3.0
matplotlib==3.8.4
numpy==1.22.0